Refactoring Cribbage game methods. Added a manager to handle turn management

+++ Deck.h
@@ -78,7 +78,7 @@ class deck {
          * @return
          */
         int getSize() {
-            return cards.size();
+            return (int) cards.size();
         }
 
         /**
